Done with Structure (?)
I think I am done with the basic structure of the game now. There are some enhancements that I could add, and I still need to add the constants (like Game board, the names, rents, and prices of properties, etc.), but I think the basic rules of the game are in place. Now I need to tie everything together.